etl/requirements.txt
osmium==3.7.0
shapely==2.0.5
geopandas==0.14.4
rasterio==1.3.10
pyproj==3.6.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.4
